+++ Classes/MapKit/MKMapView+ZoomLevel.m
@@ -148,7 +148,7 @@
 	return region;
 }
 
-- (NSUInteger) zoomLevel {
+- (double) zoomLevel {
     MKCoordinateRegion region = self.region;
     
     double centerPixelX = [MKMapView longitudeToPixelSpaceX: region.center.longitude];
